+++ src/Settings/TailwindSettings.cs
@@ -1,4 +1,6 @@
-﻿namespace TailwindCSSIntellisense.Settings
+﻿using TailwindCSSIntellisense.Options;
+
+namespace TailwindCSSIntellisense.Settings
 {
     public class TailwindSettings
     {
@@ -7,5 +9,6 @@
         public string TailwindCssFile { get; set; }
         public string TailwindOutputCssFile { get; set; }
         public bool EnableTailwindCss { get; set; }
+        public BuildProcessOptions BuildType { get; set; }
     }
 }