+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.0 (July 25th, 2023)
+
+### New Features
+
+- IntelliSense transparency support for colors (i.e. `bg-green-700/50`)
+- New build option - On Save: runs an `npx` build command on each file save; can be used if default build option is not reliable
+
+### Fixes
+
+- Fixed more classes which were not previously showing up in IntelliSense
+
 ## 1.0.3 (July 24th, 2023)
 
 ### Fixes